+++ examples/firebase-auth/src/components/SignUpForm.js
@@ -45,7 +45,9 @@ class SignUpForm extends Component {
   }
 
   render () {
-    const { username, email, passwordOne, passwordTwo, error } = this.state
+    const {
+      username, email, passwordOne, passwordTwo, error,
+    } = this.state
     const isInvalid = passwordOne !== passwordTwo ||
       passwordOne === '' || username === '' ||
       email === ''