+++ src/client/components/RouterScroller.js
@@ -5,24 +5,23 @@ import RAF from 'raf'
 import { unwrapArray } from '../../utils/shared'
 import scrollTo from '../../utils/scrollTo'
 
-const RouterScroller = withRouter(
-  class RouterScroller extends React.Component {
-    componentDidMount () {
-      this.scrollToHash()
-    }
-    componentDidUpdate (prev) {
-      if (prev.location.pathname !== this.props.location.pathname && !this.props.location.hash) {
-        if (window.__noScrollTo) {
-          window.__noScrollTo = false
-          return
-        }
-        this.scrollToTop()
+const RouterScroller = withRouter(class RouterScroller extends React.Component {
+  componentDidMount () {
+    this.scrollToHash()
+  }
+  componentDidUpdate (prev) {
+    if (prev.location.pathname !== this.props.location.pathname && !this.props.location.hash) {
+      if (window.__noScrollTo) {
+        window.__noScrollTo = false
         return
       }
-      if (prev.location.hash !== this.props.location.hash) {
-        this.scrollToHash()
-      }
+      this.scrollToTop()
+      return
     }
+    if (prev.location.hash !== this.props.location.hash) {
+      this.scrollToHash()
+    }
+  }
     scrollToTop = () => {
       const { autoScrollToTop, scrollToTopDuration } = this.props
       if (autoScrollToTop) {
@@ -72,7 +71,6 @@ const RouterScroller = withRouter(
     render () {
       return unwrapArray(this.props.children)
     }
-  }
-)
+})
 
 export default RouterScroller