+++ src/client/methods.js
@@ -32,24 +32,18 @@ export const getRouteInfo = async path => {
       try {
         if (process.env.REACT_STATIC_ENV === 'development') {
           // In dev, request from the webpack dev server
-          const { data } = await axios.get(
-            `/__react-static__/routeInfo/${path === '/' ? '' : path}`
-          )
+          const { data } = await axios.get(`/__react-static__/routeInfo/${path === '/' ? '' : path}`)
           return data
         }
         // In production, request from route's routeInfo.json
-        const { data } = await axios.get(
-          `${process.env.REACT_STATIC_PUBLIC_PATH}${pathJoin(
-            path,
-            `routeInfo.json?${process.env.REACT_STATIC_CACHE_BUST}`
-          )}`
-        )
+        const { data } = await axios.get(`${process.env.REACT_STATIC_PUBLIC_PATH}${pathJoin(
+          path,
+          `routeInfo.json?${process.env.REACT_STATIC_CACHE_BUST}`
+        )}`)
         return data
       } catch (err) {
         erroredPaths[path] = true
-        console.warn(
-          `Could not load routeInfo for path: ${originalPath}. If this is a static route, make sure any link to this page is valid! If this is not a static route, you can desregard this warning.`
-        )
+        console.warn(`Could not load routeInfo for path: ${originalPath}. If this is a static route, make sure any link to this page is valid! If this is not a static route, you can desregard this warning.`)
       }
     })()
   }
@@ -80,40 +74,35 @@ export async function prefetchData (path, { priority } = {}) {
   }
 
   // Request the template and loop over the routeInfo.sharedPropsHashes, requesting each prop
-  await Promise.all(
-    Object.keys(routeInfo.sharedPropsHashes).map(async key => {
-      const hash = routeInfo.sharedPropsHashes[key]
+  await Promise.all(Object.keys(routeInfo.sharedPropsHashes).map(async key => {
+    const hash = routeInfo.sharedPropsHashes[key]
 
-      // Check the propsByHash first
-      if (!propsByHash[hash]) {
-        // Reuse request for duplicate inflight requests
-        try {
-          if (!inflightPropHashes[hash]) {
-            if (priority) {
-              inflightPropHashes[hash] = axios.get(
-                `${process.env.REACT_STATIC_PUBLIC_PATH}staticData/${hash}.json`
-              )
-            } else {
-              inflightPropHashes[hash] = prefetchPool.add(() =>
-                axios.get(`${process.env.REACT_STATIC_PUBLIC_PATH}staticData/${hash}.json`)
-              )
-            }
+    // Check the propsByHash first
+    if (!propsByHash[hash]) {
+      // Reuse request for duplicate inflight requests
+      try {
+        if (!inflightPropHashes[hash]) {
+          if (priority) {
+            inflightPropHashes[hash] = axios.get(`${process.env.REACT_STATIC_PUBLIC_PATH}staticData/${hash}.json`)
+          } else {
+            inflightPropHashes[hash] = prefetchPool.add(() =>
+              axios.get(`${process.env.REACT_STATIC_PUBLIC_PATH}staticData/${hash}.json`))
           }
-          const { data: prop } = await inflightPropHashes[hash]
-
-          // Place it in the cache
-          propsByHash[hash] = prop
-        } catch (err) {
-          console.error('Error: There was an error retrieving a prop for this route! hashID:', hash)
-          console.error(err)
         }
-        delete inflightPropHashes[hash]
-      }
+        const { data: prop } = await inflightPropHashes[hash]
 
-      // Otherwise, just set it as the key
-      allProps[key] = propsByHash[hash]
-    })
-  )
+        // Place it in the cache
+        propsByHash[hash] = prop
+      } catch (err) {
+        console.error('Error: There was an error retrieving a prop for this route! hashID:', hash)
+        console.error(err)
+      }
+      delete inflightPropHashes[hash]
+    }
+
+    // Otherwise, just set it as the key
+    allProps[key] = propsByHash[hash]
+  }))
 
   // Cache the entire props for the route
   routeInfo.allProps = allProps