+++ src/static/webpack/index.js
@@ -63,13 +63,7 @@ export async function startDevServer ({ config }) {
   const intendedPort = (config.devServer && config.devServer.port) || process.env.PORT || 3000
   const port = await findAvailablePort(Number(intendedPort))
   if (intendedPort !== port) {
-    console.time(
-      chalk.red(
-        `=> Warning! Port ${intendedPort} is not available. Using port ${chalk.green(
-          intendedPort
-        )} instead!`
-      )
-    )
+    console.time(chalk.red(`=> Warning! Port ${intendedPort} is not available. Using port ${chalk.green(intendedPort)} instead!`))
   }
   const host = (config.devServer && config.devServer.host) || process.env.HOST || 'http://localhost'
 
@@ -212,31 +206,25 @@ export async function buildProductionBundles ({ config }) {
         const buildWarnings = stageStats.hasWarnings()
 
         if (buildErrors || buildWarnings) {
-          console.log(
-            stageStats.toString({
-              context: config.context,
-              performance: false,
-              hash: false,
-              timings: true,
-              entrypoints: false,
-              chunkOrigins: false,
-              chunkModules: false,
-              colors: true,
-            })
-          )
+          console.log(stageStats.toString({
+            context: config.context,
+            performance: false,
+            hash: false,
+            timings: true,
+            entrypoints: false,
+            chunkOrigins: false,
+            chunkModules: false,
+            colors: true,
+          }))
           if (buildErrors) {
-            console.log(
-              chalk.red.bold(`
+            console.log(chalk.red.bold(`
                 => There were ERRORS during the ${stage} build stage! :(
                 => Fix them and try again!
-              `)
-            )
+              `))
           } else if (buildWarnings) {
-            console.log(
-              chalk.yellow.bold(`
+            console.log(chalk.yellow.bold(`
                 => There were WARNINGS during the ${stage} build stage!
-              `)
-            )
+              `))
           }
         }
       }