+++ src/utils/shared.js
@@ -42,9 +42,7 @@ export function isObject (a) {
 }
 
 export function deprecate (from, to) {
-  console.warn(
-    `React-Static deprecation notice: ${from} will be deprecated in favor of ${to} in the next major release.`
-  )
+  console.warn(`React-Static deprecation notice: ${from} will be deprecated in favor of ${to} in the next major release.`)
 }
 
 export function trimSlashes (str) {